requirements.txt
streamlit
Pillow
torch
torchvision
torchmetrics
numpy
